+++ packages/analytics-js/.size-limit.js
@@ -19,13 +19,13 @@ module.exports = [
     name: 'Core - NPM',
     path: 'dist/npm/modern/umd/index.js',
     gzip: true,
-    limit: '24 KiB',
+    limit: '24.5 KiB',
   },
   {
     name: 'Core Legacy - CDN',
     path: 'dist/cdn/legacy/iife/rsa.min.js',
     gzip: true,
-    limit: '47.5 KiB',
+    limit: '48.5 KiB',
   },
   {
     name: 'Core - CDN',