+++ packages/analytics-js/src/components/core/Analytics.ts
@@ -325,7 +325,9 @@ class Analytics implements IAnalytics {
     // as this will prevent us from supporting multiple SDK instances in the same page
     // Execute onLoaded callback if provided in load options
     if (isFunction(state.loadOptions.value.onLoaded)) {
-      (state.loadOptions.value.onLoaded as OnLoadedCallback)((globalThis as any).rudderanalytics);
+      (state.loadOptions.value.onLoaded as OnLoadedCallback)(
+        (globalThis as typeof window).rudderanalytics,
+      );
     }
 
     // Set lifecycle state
@@ -338,7 +340,7 @@ class Analytics implements IAnalytics {
 
     // Emit an event to use as substitute to the onLoaded callback
     const initializedEvent = new CustomEvent('RSA_Initialised', {
-      detail: { analyticsInstance: (globalThis as any).rudderanalytics },
+      detail: { analyticsInstance: (globalThis as typeof window).rudderanalytics },
       bubbles: true,
       cancelable: true,
       composed: true,
@@ -354,7 +356,7 @@ class Analytics implements IAnalytics {
   onReady() {
     // Emit an event to use as substitute to the ready callback
     const readyEvent = new CustomEvent('RSA_Ready', {
-      detail: { analyticsInstance: (globalThis as any).rudderanalytics },
+      detail: { analyticsInstance: (globalThis as typeof window).rudderanalytics },
       bubbles: true,
       cancelable: true,
       composed: true,
@@ -691,6 +693,10 @@ class Analytics implements IAnalytics {
   getSessionId(): Nullable<number> {
     const sessionId = this.userSessionManager?.getSessionId();
     return sessionId ?? null;
+  }
+
+  setAuthToken(token: string): void {
+    this.userSessionManager?.setAuthToken(token);
   }
   // End consumer exposed methods
 }