+++ packages/analytics-js/src/components/userSessionManager/UserSessionManager.ts
@@ -19,10 +19,18 @@ import { ApiObject } from '@rudderstack/analytics-js-common/types/ApiObject';
 import { AnonymousIdOptions } from '@rudderstack/analytics-js-common/types/LoadOptions';
 import { USER_SESSION_MANAGER } from '@rudderstack/analytics-js-common/constants/loggerContexts';
 import { StorageType } from '@rudderstack/analytics-js-common/types/Storage';
-import { COOKIE_STORAGE, LOCAL_STORAGE } from '@rudderstack/analytics-js-common/constants/storages';
+import {
+  COOKIE_STORAGE,
+  LOCAL_STORAGE,
+  SESSION_STORAGE,
+} from '@rudderstack/analytics-js-common/constants/storages';
 import { UserSessionKeys } from '@rudderstack/analytics-js-common/types/userSessionStorageKeys';
 import { StorageEntries } from '@rudderstack/analytics-js-common/types/ApplicationState';
-import { CLIENT_DATA_STORE_COOKIE, CLIENT_DATA_STORE_LS } from '../../constants/storage';
+import {
+  CLIENT_DATA_STORE_COOKIE,
+  CLIENT_DATA_STORE_LS,
+  CLIENT_DATA_STORE_SESSION,
+} from '../../constants/storage';
 import { storageClientDataStoreNameMap } from '../../services/StoreManager/types';
 import { DEFAULT_SESSION_TIMEOUT_MS, MIN_SESSION_TIMEOUT_MS } from '../../constants/timeouts';
 import { defaultSessionInfo } from '../../state/slices/session';
@@ -76,7 +84,7 @@ class UserSessionManager implements IUserSessionManager {
     const userTraits = this.getUserTraits();
     const groupId = this.getGroupId();
     const groupTraits = this.getGroupTraits();
-    const anonymousId = this.getAnonymousId();
+    const anonymousId = this.getAnonymousId(state.loadOptions.value.anonymousIdOptions);
     if (userId) {
       this.setUserId(userId);
     }
@@ -91,6 +99,10 @@ class UserSessionManager implements IUserSessionManager {
     }
     if (anonymousId) {
       this.setAnonymousId(anonymousId);
+    }
+    const authToken = this.getAuthToken();
+    if (authToken) {
+      this.setAuthToken(authToken);
     }
 
     const initialReferrer = this.getInitialReferrer();
@@ -126,7 +138,7 @@ class UserSessionManager implements IUserSessionManager {
       const key = entry as UserSessionStorageKeysType;
       const currentStorage = entries[key]?.type as StorageType;
       const curStore = this.storeManager?.getStore(storageClientDataStoreNameMap[currentStorage]);
-      const storages = [COOKIE_STORAGE, LOCAL_STORAGE];
+      const storages = [COOKIE_STORAGE, LOCAL_STORAGE, SESSION_STORAGE];
 
       storages.forEach(storage => {
         const store = this.storeManager?.getStore(storageClientDataStoreNameMap[storage]);
@@ -149,12 +161,16 @@ class UserSessionManager implements IUserSessionManager {
     }
     const cookieStorage = this.storeManager?.getStore(CLIENT_DATA_STORE_COOKIE);
     const localStorage = this.storeManager?.getStore(CLIENT_DATA_STORE_LS);
+    const sessionStorage = this.storeManager?.getStore(CLIENT_DATA_STORE_SESSION);
     const stores: IStore[] = [];
     if (cookieStorage) {
       stores.push(cookieStorage);
     }
     if (localStorage) {
       stores.push(localStorage);
+    }
+    if (sessionStorage) {
+      stores.push(sessionStorage);
     }
     Object.keys(userSessionStorageKeys).forEach(storageEntryKey => {
       const key = storageEntryKey as UserSessionStorageKeysType;
@@ -311,6 +327,12 @@ class UserSessionManager implements IUserSessionManager {
     effect(() => {
       this.syncValueToStorage('sessionInfo', state.session.sessionInfo.value);
     });
+    /**
+     * Update session tracking info in storage automatically when it is updated in state
+     */
+    effect(() => {
+      this.syncValueToStorage('authToken', state.session.authToken.value);
+    });
   }
 
   /**
@@ -323,8 +345,7 @@ class UserSessionManager implements IUserSessionManager {
    */
   setAnonymousId(anonymousId?: string, rudderAmpLinkerParam?: string) {
     let finalAnonymousId: string | undefined | null = anonymousId;
-    const storage: StorageType = state.storage.entries.value.anonymousId?.type as StorageType;
-    if (isStorageTypeValidForStoringData(storage)) {
+    if (this.isPersistenceEnabledForStorageEntry('anonymousId')) {
       if (!finalAnonymousId && rudderAmpLinkerParam) {
         const linkerPluginsResult = this.pluginsManager?.invokeMultiple<Nullable<string>>(
           'userSession.anonymousIdGoogleLinker',
@@ -439,6 +460,14 @@ class UserSessionManager implements IUserSessionManager {
   }
 
   /**
+   * Fetches auth token from storage
+   * @returns
+   */
+  getAuthToken(): Nullable<string> {
+    return this.getItem('authToken');
+  }
+
+  /**
    * If session is active it returns the sessionId
    * @returns
    */
@@ -489,9 +518,11 @@ class UserSessionManager implements IUserSessionManager {
       state.session.userTraits.value = defaultUserSessionValues.userTraits;
       state.session.groupId.value = defaultUserSessionValues.groupId;
       state.session.groupTraits.value = defaultUserSessionValues.groupTraits;
+      state.session.authToken.value = defaultUserSessionValues.authToken;
 
       if (resetAnonymousId) {
-        state.session.anonymousId.value = defaultUserSessionValues.anonymousId;
+        // This will generate a new anonymous ID
+        this.setAnonymousId();
       }
 
       if (noNewSessionStart) {
@@ -608,6 +639,16 @@ class UserSessionManager implements IUserSessionManager {
    */
   end() {
     state.session.sessionInfo.value = {};
+  }
+
+  /**
+   * Set auth token
+   * @param userId
+   */
+  setAuthToken(token: string) {
+    if (this.isPersistenceEnabledForStorageEntry('authToken')) {
+      state.session.authToken.value = token;
+    }
   }
 }