+++ packages/analytics-js/src/index.ts
@@ -1,4 +1,7 @@
-// TODO: export all types that we need for public interface consumption
+import { RudderAnalytics } from './app/RudderAnalytics';
+import { PreloadedEventCall } from './components/preloadBuffer/types';
+import { IRudderStackGlobals } from './app/IRudderStackGlobals';
+
 export {
   type AnonymousIdOptions,
   type LoadOptions,
@@ -18,3 +21,12 @@ export { type PluginName } from '@rudderstack/analytics-js-common/types/PluginsM
 export { RudderAnalytics } from './app/RudderAnalytics';
 export { type IRudderStackGlobals } from './app/IRudderStackGlobals';
 export { type PreloadedEventCall } from './components/preloadBuffer/types';
+
+declare global {
+  interface Window {
+    rudderanalytics: RudderAnalytics | PreloadedEventCall[];
+    RudderStackGlobals: IRudderStackGlobals;
+    rudderAnalyticsMount: () => void;
+    rudderAnalyticsBuildType: 'legacy' | 'modern';
+  }
+}