+++ packages/sanity-suite/src/ignoredProperties/sourceConfigIgnoredProperties.js
@@ -1,0 +1,100 @@
+const sourceConfigIgnoredProperties = [
+  {
+    key: `updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.liveEventsConfig.eventUploadTS`,
+    type: 'number',
+  },
+  {
+    key: `source.sourceDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.0.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.0.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.0.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.1.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.1.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.1.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.2.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.2.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.2.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.3.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.3.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.3.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.3.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.4.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.4.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.5.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.5.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.5.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.6.updatedAt`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.6.revisionId`,
+    type: 'string',
+  },
+  {
+    key: `source.destinations.6.destinationDefinition.updatedAt`,
+    type: 'string',
+  },
+];
+
+export { sourceConfigIgnoredProperties };